Martin DeMello's patch to add dxf import code for: CIRCLE, ARC, ELLIPSE and LWPOLYLINE.
I made some modifications to it to clean it up a bit.

I entend to do some more dxf stuff in  the near future but figured
this is ready to go for now, and trying to clean up a bunch of small stuff
I've had sitting around for awhile. ;)

Kent