Bug #1865
On appending in posemoding, Blender could crash.
Found out the new 'select appended' also sets the 'active base/object'.
This isn't OK on that level of the code, 'active' exists (and needs to be
set) on UI code level in src/, not in the middle of file reading function.

Also removed weird negative bitflag that enforced 'select appended' to
be default. If you want that, set the flag itself in do_versions or so.
In my opinion it is best saved in .B.blend instead. Another time...