Hrms... the fix for bug #4010 didn't work as expected.
The issue was that particle emittors were still transformed by the object
matrix itself. That was solved in the previous commit, but there was
still an error in correctly evaluating dependencies for the object...
Current commit uses depsgraph to recalculate all objects that influence
the emittor.

The depsgraph code doesn't like particles much (because it uses baking).
Current construct is still weak, is on the list to solve nice.