Bugfix 4865
Manipulator orientation tweak: when having multiple objects selected, the
orientation option "Local" was always skipped, showing it global. Now it
uses the orientation of the active object.

The bug reportor noticed a conflict with the "Around" option "Active". In
that case he expected the orientation to be from Active Object as well.