Plumifero's wishlist:
* When duplicating bones that have constraints (edit mode),
duplicate them too with TARGET field updated.

This is code that I wrote a few years ago before the armature
refactor. It is simple code that compiles and seems to work fine,
but should be checked by Ton or somebody else with better knowledge
of the current armature system.

Usage description, (stolen from the first time this code was
committed):

"Duplicating bones in edit mode now also duplicates the constraints
associated with that bone ... if the constraint subtarget bone is
also duplicated the new constraint points to this new bone as it's
subtarget."