== PoseLib - Quick Search for Preview ==
Now, when Previewing/Browsing poses with Ctrl-L, it is possible to type in part of a name to limit the poses which can get chosen to hone in on the poses you wish to apply more quickly. 
* The search is case insensitive, and doesn't look for complete matches (it will 'match' if the whole search-string is found). 
* Only basic text manipulation capabilities are available for editing the search string (i.e.  backspace, del, home, end, move text-cursor forward ->, move text-cursor backwards <-).
* Pose browsing using the Scroll-Wheel and Page up/down can be used to browse through the matches
* View manipulation is only possible with MMB, as numpad keys are used for search-string input.

In the process, I've recoded the preview function to be more modular. It's now split up more, like the Transform core.