Fix segfault when save the render with Multilayer format.
The RE_WriteResult function don't call BLI_make_existing_file
like BKE_write_ibuf, so if you append a directory to the output
path and it don't exist, the exr_begin_write fail to create
the file and blender crash.

Note that this is only 1 line of code and don't touch
anything in the render code.