CMake files update to allow use of MAC OSX 10.5 libs (and build 64bit blender on mac)
The WITH_LIBS10.5 option switches the use of the libs included in the darwin-9.x.universal folder

Use the CMAKE_OSX_ARCHITECTURES variable to set the architecture you want to build for (e.g. i386, x86_64). Only one at a time, this value is used to select the python_?.zip that is bundled with the app.

WITH_COCOA (build Cocoa ghost and not Carbon) is now on by default.