Added RNA based transform properties for pose bones too.
I've left the old code commented out for now hough - do we still want to 
be able to edit all rotation types as eulers in the transform properties buttons? 
Seems a bit odd to me, what do animators think? If so, maybe we need some 
ui-level conversion options in the RNA buttons code...