BGE: rna update for Logic Bricks
New Actuators done:
* EDIT_OBJECT
* SHAPE_ACTION

Actuators to be done:
* CONSTRAINT

Actuators done already:
* ALL THE OTHERS 16

+ some DNA padding

There is one Actuator left. This one is tricky because (as many others) the tooltip changes according to the actuator type value. Not sure how to solve this yet. I'll probably have to make different rna properties for the same values (it includes some enums this time). Not so animated to finish this one :/