Bugfix #26812
On anim-render, a click in timeline stopped render completely.

The reason for this was a bit wacko code to cope with frame-step 
feature (steps of multiple frames).

I thought of fixing that, but instead decided to block any operator 
in Blender to change a frame while a render is in progress. 
Both render engine and UI are accessing (writing to) the same 
data then, which is a bad conflict.

Still a serious weakness of threaded render, but I'll keep
trying to allow this as far as possible :)