Turned select vertex path back on and added back in selection history for bmesh->mesh and mesh->bmesh conversions: ---------------------------------------------- Select Vertex Path had wrong invoke callback (shouldn't have any?).
Also selection history was not converted when 
doing bmesh->mesh or mesh->bmesh conversions. 
This meant that tabbing in and out of editmode
would make your selection history dissapear.
Undo pop would also not preserve selection 
history so any operators that relied on it
would not work when you adjusted their settings.