Fix 30169: rotate edge needs more checking for bad cases
If you tried to rotate an edge that was part of a chain
between two faces, the join faces function would remove
the whole chain (else it would leave a 'spur') and then
rotate edge would crash looking for a vertex that is no
longer there. 
Check for this case and disallow rotate if so.