Bugfix:
While testing the other my previous commit (removing deformflag stuff), I
noticed that trying to change set a parent object for the current object (via
the Object Properties) would cause the current object to "blow up", and for
cyclic dependency warnings to be spewed to the console.

Adding a dummy usage of one of the vars here seems to solve it. Perhaps a case
of weirdo compiler optimisations?