Tomato Cycles: do not discard unfinished tiles on cancel
Do not discard (fill with black) tiles which are not fully rendered
(not all the samples are calculated for tile) when canceling render.

This could be helpful to tweak some settings when render glitch is
discovered. Also it could be used in such scenarios as setting
samples number to something really high and render still image
until result is reasonable, controlling this manually.

This could make cancel not so responsible on CPU, but it wouldn't
be less responsible than GPU, also could potentially give some
%% of speedup by avoiding checking cancel state after every pixel
sampled.