Color Management: added check for invertibility of color spaces
So now non-invertible color spaces can not be used as input color space
for images and as working color space for sequencer.

Currently have got two hard-coded families which are rrt and display.
If color space belong to one of this families, it would be considered
as non-invertible.

Data color spaces are always considered invertible.

If color space has got to_reference transformation, it'll be considered
invertible.