Bugfix [#32501] Protected layer bones transform buttons are greyed out/inactive
Since the transform values for bones in protected layers can be animated again
using many of the common tools, it makes sense that the transform properties for
these are allowed to be edited too.

I've left the rotation mode setting as a "proxy locked" value for now, since
this seems more like something that's defined as part of the rig creation (and
best left alone by animators).