rigidbody: Add "Apply Transformation" button
This is just the "Apply Visual Transform" operator.
It's very usefull for rigid body simulations but hard to find and users
usually don't know about it/don't know it's usefull to apply rigid body
transformations.
It seems bit out of place (especially the tooltip) so we might need to
do a bit more here.