Initialize git submodules for addons, locales and scons
Uses relative paths to repositories, so this is expected
to work fine for any protocol we support (git, ssh and http).

Uses ignore=all for all the submodules, so updating them
to latest remote hash does not tags blender repository
as changes. But it is still possible to make changes to
submodules and commit them from their path.