Object Mode Toolbar Tabs
Summary:
Here are the initial tabs for the 3D View Object Mode toolbar. This is based on the discussions between myself and @billrey in T37569

P.S. apologies for the duplicate. The first one I mistakenly sent to the Edit Mode task, T37568. I have corrected it and also made a couple changes, per @campbellbarton's request.

Reviewers: campbellbarton

CC: campbellbarton, brecht, billrey

Maniphest Tasks: T37569

Differential Revision: http://developer.blender.org/D129