Revert "Fix off by one error in display of start/end frame in sequencer."
This reverts commit 0e02ad8b6495b2755598a24b487041a3ed8e0116.

Initial commit was done so visual result fits with animation cursor
in timeline but this makes it so it looks like one extra frame is
rendered. Other idea would be to render one less frame for sequencer
but this is not so nice either. Generally here's no way to be
fully consistent here, but at least let's be workflow-consistent