Fix T62841: Hidden collection shortcut should be work in edit mode
This is a full revert of b6c61945aed51d86dc474df616761f9a25d71d09.

But ammended to include the object.hide_collection() operator to the
pose keymap as well.

While trying to expose the shortcuts to pose mode, I ended up exposing
them for all the modes. It is now only available for object and mode
modes.