Sculpt dyntopo: Wrote a little UV smoother sculpt brush. It uses a simple LSCM-like solver; hold ctrl to do simple laplacian relaxation.
There are some interesting artistic possibilities to using
DynTopo with UV maps, the point of this tool is to help
explore them.

Note that I'm not planning to merge this into master with
the rest of this branch.  When the time comes I plan to
move it into a new branch for sculpt research stuff.