Sculpt-dev: commit 1 of dyntopo cleanup
In an attempt to fix dyntopo artifacts with
snake hook I wrote a new heap implementation
for dyntopo, a min-max heap.

Turns out the seperation of collapse and
subdivision queues was a fundamental design
flaw in DynTopo.  Unifiying them in
a single heap is much faster (as much as
2x faster depending on the setup), and
will allow for much cleaner code.