XR: Enable (selectively) Vive Tracker extension
Allows action bindings for the HTC Vive Trackers, so that their poses
can be queried and used for motion capture objects. For details on the
available tracker roles/user path identifiers, see: https://www.khronos.org/registry/OpenXR/specs/1.0/html/xrspec.html#XR_HTCX_vive_tracker_interaction

Currently the extension is only supported by the SteamVR runtime (1.21)
and there a few issues (which could be on the Blender or SteamVR side),
such as incorrect controller poses when the extension is enabled.

For this reason, the tracker extension is added as a session option to
enable only when one wants to exclusively use trackers instead of
controllers.