Fix: Incorrectly sized curves created in set conversion node
The number of points in the source curve was needed, but the offset
(just zero) was passed instead. It's unclear how this worked before.
A mistake in the recent commit 9e393fc2f12583d32dd.

Also use a common utility for retrieving the sizes of curves
in ranges instead of reimplementing it for this file.