USD: New export UI.
Change authored by Charles Wardlaw.

UI for export is now split into multiple collapsible tool
panels on the right side.  Added a function to io_ops.c to
facilitate this, as there is no current API for this
(checked with the BF).

As per discussions, defaults have also been updated and
some options, while remaining available through the
Python API for the operator, are now removed from the
Export list.