USD: panels UX change for import/export options.
Operator UX adjusted to group functionality into panels
for groupings of related settings.

Some defaults have been adjusted.  Not all options in the
Operators may be visible in the new UI, instead allowing
their defaults to become the only behavior. However, those
options are not removed from the Operators so that script
authors can take advantage of the extra flexibility.