USD import: fix missing packed textures for UMM.
The material importer now always attempts to pack textures
of any texture image node that were created, even if the UMM
conversion was only partially complete, since it is useful
to render a parially-constructed material.  For example,
if the material has a valid color texture but the
normal texture is missing, we still want to load the color
texture in the viewport, regardless of any UMM error due
to the missing normal texture.