temp-sculpt-dyntopo: Make use of DynTopoSettings again
All DynTopo settings are now stored in a single struct,
DynTopoSettings, which is used by both Brush and Sculpt.

DynTopo settings can be set to either inherit from Sculpt.dyntopo
or Brush.dyntopo; this in controlled with Brush.dyntopo.inherit
which is a bitflag.

This is the original system I came up prior to the more
generic brush property proposals.  I decided to bring it
back for the same reason I wrote it in the first place:
to avoid hardcoding behavior for specific brushes, specifically:

* Simplify always has subdivide and collapse on.
* Snake hook needs a much smaller dyntopo
  spacing than other tools.