temp-sculpt-dyntopo: Remove pattern based triangle subdivider
This turned out to be a dead end. The original idea of using a
pattern based subdivider was to preserve quads.  But that turned
out to be more trouble than it was worth and just converged to
triangle in the end anyway, except for little floating quads
that messed up smoothing.

So I removed support for quads (quite a while ago) and recent
tests have shown a simple edge splitter is faster and produces
higher quality results.