Fix: Crash when using loop select on edge that has only 1 face
The code assumed that the edge to poly map returns 2 polys,
which of course is not always the case, e.g. the border edges on a plane

Now the code assumes there is at least 1 face
which should be true since the code picks a face to begin with.

Pull Request: https://projects.blender.org/blender/blender/pulls/108759