I18n: use proper name for the Convert Colorspace node
The Convert Colorspace node uses "Convert Colorspace" as a name on
registration, but it uses "Color Space" as a `DefNode()` in
NOD_static_types.h.

As a result, the node name is not extracted to the .po files, and thus
not translated.

This commit uses the same name in all definitions of the node.

Pull Request: https://projects.blender.org/blender/blender/pulls/109419