temp-sculpt-dyntopo: Fix uv collapse bug
Apparently you have to use minmax instead of centroid
when collapsing UV edges.  I thought I had all the
weights summing to produce the same result, but apparently
not.  This is also how editmode does it btw, that's where
I finally found the solution.

In other news, there is now a debug option to show UV edges
in the uv editor in sculpt mode (in dynamic topology mode).
It's purely for testing and is not meant for master.
To use, simply uncomment the DEBUG_SHOW_SCULPT_BM_UV_EDGES definition
in BKE_sculpt.h.