Make asset poll callback non-temporary design
While this was initially meant as temporary design, I think it's fine to
keep this. It adds more flexibility in removing unwanted assets from the
listing. Plus it's unclear if we can get the necessary Python API syntax
to work before the release, so I prefer not relying on that. Should we
get it to work, the API can still be changed before the release.