UI: Theme Editor: Mention editors that share settings
It's not immediately clear for new users that editing the
"File Browser" colors will also adjust the "Asset Browser".

Similar to how it's already the case for UV/Image Editor,
mention in the name of the panel other editors that share
the same color settings.

* Dope Sheet > Dope Sheet/Timeline
* File Browser > File/Asset Browser
* Graph Editor > Graph Editor/Drivers

The only exception is Node Editor because eventually it
will get too long.