Refactor: Code indentation in transform_convert_nla.cc
No functional changes

`transform_convert_nla.cc` had a few `if`s within
for loops that could be inverted to remove the indentation.

In addition to that I extracted the snapping functionality
so the `if`s there can also be inverted and return.

Pull Request: https://projects.blender.org/blender/blender/pulls/111968