Show FCurve modifier toggle widget as 'inactive' when there are no modifiers
In the animation editors, show the FCurve Modifier toggle widget (the
wrench) as 'inactive' when the FCurve has no modifiers on it. Previously
the (lack of) existence of FCurve modifiers was not reflected in the
wrench icon at all.

Note: this only affects FCurves, and not FCurve *Groups*.

Pull Request: https://projects.blender.org/blender/blender/pulls/106214