Animation: Graph Editor locked key drawing
As people have pointed out, the changes to key drawing
in #106052 are not ideal under all conditions.

The fact that an X is drawn to indicate a key
can't be manipulated adds too much noise
on dense data and is generally confusing.

To combat that, the drawing style changed it back to dots,
but this time in gray ( the same color as disabled curves)
and slightly smaller.

This was discussed in the A&R module meeting on 2023-09-07

Pull Request: https://projects.blender.org/blender/blender/pulls/111986