UI: Improve/correct vertical widget positioning in header-like regions
There was code to vertically offset layouts in header-like regions to
ensure widgets are centered properly. It wasn't clear why this was
necessary so far. I noticed this is region alignment dependent though:
The offset is only needed when the upper region edge matches the upper
area edge. I'm not entirely sure if it is the edge drawing or a slightly
smaller drawable surface of the region (clipped to be 1px less than the
area height), think it is the latter. But either way, this offsetting
makes a lot more sense now. Fix is to only apply the offset for regions
sharing the upper edge with the area. Edges at the window differ a bit,
so the offset is not applied there.

Fixes the slightly off vertical positioning of widgets in the asset
shelf header.