Fix #112123: only some vertex weights editable (Blender 3.6)
I couldn't reproduce the issue myself, but my guess is that this happened to work for me because freeing the pointer and then immediately allocating the same size again yielded the same pointer. This is allocator dependent behavior which depends on the platform.

A stable pointer is necessary so that `uiBut`s of different frames can be identified to correspond to each other.

In `main` that works already, because `Vector.resize` does not change the pointer if the size does not change.

Pull Request: https://projects.blender.org/blender/blender/pulls/112360