USD export: skip shader inputs.
Now skipping the "Factor" input when traversing
channels.  This conditional can be extended to exclude
other input names in the future.  This heuristic
helps skip inputs that aren't likely to be suitable
as UsdPreviewSurface input values.  For example,
it would prevent exporting an SSS texture that is
connected to the Factor input of a Mix node.