Fix #115715: Boolean behaves as if the object is on it's origin
Resolve regression in [0] which replaced a null check with an
identity comparison. This wasn't correct as the null check indicated
that transform was applied or not, where situations that require a
transform could have an identity matrix.

Resolve by checking if the matrix that would transform vertices is an
identity which is a reliable way to know if any transformation is needed.

Ref: !115736

[0]: e24c7f1954b7ac3e40fda1172789d118c43fa6f1

---

Back ported from main, c9dd5ce4cc1f2f8ed6a20e92d8778ef985db0aa6