Fix #117900: Sculpt mode doesn't clear face corner normals
In sculpt mode vertex and face normals are updated explicitly, but
corner normals, which are cached now, aren't updated for performance
reasons. This means that when the mesh is deformed, they need to be
cleared. See b339e3937d8dc580b655 for some background on the complexity
of this area.