Brush Assets: Add library choice to "Save As" operator
Previously the new brush would always be saved to the default library.
Now the user is presented with a choice that defaults to the library marked
as "Default" in the preferences.

Because the library is now chosen by an operator property, we can't check
its editable status in the poll function. Instead we just check it when the
operator executes.

Pull Request: https://projects.blender.org/blender/blender/pulls/118021