Make asset shelf visible by default in sculpt/paint modes
Adds a new `DEFAULT_VISIBLE` option for the `bl_options` of an asset
shelf. When this asset shelf will be created for the first time, the
asset shelf will be made visible, otherwise it will be hidden.

We don't mess with the visibility further, meaning, when re-opening a
previously active asset shelf, we keep the current asset shelf
visibility. So only the first load is affected. Alternatively asset
shelves could remember their last visibility state, and restore that as
context (e.g. the mode) changes. To be evaluated.

Had to add some trickery to make this overriding of visibility possible.