Brush Assets: Add catalog option to asset pushing operator
Adds the ability to choose which of the libraries catalogs to add the
new asset to, or to create a new catalog. The catalogs are presented
in a dropdown list, and every hierarchy level is available. Eventually
we might want some form of tree view here. But for now this approach is
very simple.

Pull Request: https://projects.blender.org/blender/blender/pulls/118382