USD: texture IO improvements
Added a "get_image_path_cb" callback as an argument to the UMM
export_material function, to allow resolving texture asset paths
from UMM based on the current export parameters.

Now recording the original USD asset path of textures imported from
URIs in a 'usd_source_path' custom propery on the imported Blender
image.

Added a 'use_original_paths' USD export option which, if enabled,
will cause the oiginal URI saved in an image's 'usd_source_path'
custom property to exported as the asset path for that texture
in the USD.  This option is not available when exporting textures.

Fixed relative path computation for texture URIs.

Fixed error loading UDIM textures imported from UMM by setting
missing tile information on the imported images.

Fixed logic error which was causing UDIM textures to be packed in the
Blender file even when the textures are on the local file system and
don't need to be imported.