USD: fix crashes exporting animating instances
Addressed two issues which were causing exceptions when
exporting animating instances:

Fixed issue causing prototypes to be incorrectly marked
as intances in some cases when exporting animations.

Fixed issue which was causing more than one instance
to be marked as "the original" when setting duplication
references when exporting animations.  This was causing
USD errors and exceptions because of circular references.