USD export: set displayName for non-ASCII names
For users writing object names in non-ASCII characters, this patch
aims to more automatically handle the creation of displayName
metadata at export time. If an Object, Mesh Data, or Material name
contains non-ASCII characters, instead of converting them all to
underscores (which can lead to prim names made entirely of
underscores, and cause name collisions), a new unique name is
generated and the original name is assigned to the prim's displayName.

Note: if a string Custom Data property called "displayName" is
assigned to the object, this takes priority! This patch is meant to
extend the prior patch on displayNames, not override it. The
automatic names are only assigned if the name contains non-ASCII
characters and there is not already a displayName set by the user
on the Blender side.